requirements.txt
# Core LLM / Agent framework
langchain>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.2.0,<0.3.0

# API framework
fastapi>=0.115.0,<0.120.0
uvicorn[standard]>=0.30.0,<0.35.0
python-multipart>=0.0.12

# Dashboard
streamlit>=1.40.0,<2.0.0
plotly>=5.24.0,<6.0.0

# Data processing
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.1.0

# HTTP client (for Streamlit → FastAPI calls)
requests>=2.32.0,<3.0.0
httpx>=0.27.0,<0.30.0

# Environment / config
python-dotenv>=1.0.0,<2.0.0

# Validation
pydantic>=2.9.0,<3.0.0
